Update to 2.8.4 step-1

+++ application/controllers/admin/LegalServices/Cases_controller.php
@@ -23,7 +23,7 @@ class Cases_controller extends AdminController
 
     public function add($ServID)
     {
-        if (!has_permission('projects', '', 'edit') && !has_permission('projects', '', 'create')) {
+        if (!staff_can('edit', 'projects') && !staff_can('create', 'projects')) {
             access_denied('Projects');
         }
         $ExistServ = $this->legal->CheckExistService($ServID);
@@ -44,6 +44,10 @@ class Cases_controller extends AdminController
         $data['service'] = $this->legal->get_service_by_id($ServID)->row();
         $data['Numbering'] = $this->case->GetTopNumbering();
         $data['auto_select_billing_type'] = $this->case->get_most_used_billing_type();
+        if ($this->input->get('via_estimate_id')) {
+            $this->load->model('estimates_model');
+            $data['estimate'] = $this->estimates_model->get($this->input->get('via_estimate_id'));
+        }
         $data['last_case_settings'] = $this->case->get_last_case_settings();
         if (count($data['last_case_settings'])) {
             $key                                       = array_search('available_features', array_column($data['last_case_settings'], 'name'));
@@ -62,7 +66,7 @@ class Cases_controller extends AdminController
 
     public function edit($ServID,$id)
     {
-        if (!has_permission('projects', '', 'edit') && !has_permission('projects', '', 'create')) {
+        if (!staff_can('edit', 'projects') && !staff_can('create', 'projects')) {
             access_denied('Projects');
         }
         if(!$id){
@@ -88,6 +92,10 @@ class Cases_controller extends AdminController
         $data['case'] = $this->case->get($id);
         $data['case']->settings->available_features = unserialize($data['case']->settings->available_features);
         $data['auto_select_billing_type'] = $this->case->get_most_used_billing_type();
+        if ($this->input->get('via_estimate_id')) {
+            $this->load->model('estimates_model');
+            $data['estimate'] = $this->estimates_model->get($this->input->get('via_estimate_id'));
+        }
         $data['last_case_settings'] = $this->case->get_last_case_settings();
         if (count($data['last_case_settings'])) {
             $key                                          = array_search('available_features', array_column($data['last_case_settings'], 'name'));
@@ -106,24 +114,26 @@ class Cases_controller extends AdminController
 
     public function delete($ServID,$id)
     {
-        if (!has_permission('legal_recycle_bin', '', 'delete')) {
-            access_denied('legal_recycle_bin');
-        }
-        if(!$id){
-            set_alert('danger', _l('WrongEntry'));
+        if (staff_can('delete', 'projects')) {
+            if(!$id){
+                set_alert('danger', _l('WrongEntry'));
+                redirect(admin_url("LegalServices/LegalServices_controller/legal_recycle_bin/$ServID"));
+            }
+            $response = $this->case->delete($ServID,$id);
+            if ($response == true) {
+                set_alert('success', _l('deleted_successfully'));
+            } else {
+                set_alert('warning', _l('problem_deleting'));
+            }
             redirect(admin_url("LegalServices/LegalServices_controller/legal_recycle_bin/$ServID"));
         }
-        $response = $this->case->delete($ServID,$id);
-        if ($response == true) {
-            set_alert('success', _l('deleted_successfully'));
-        } else {
-            set_alert('warning', _l('problem_deleting'));
-        }
-        redirect(admin_url("LegalServices/LegalServices_controller/legal_recycle_bin/$ServID"));
     }
 
     public function move_to_recycle_bin($ServID,$id)
     {
+        if (!has_permission('legal_recycle_bin', '', 'delete')) {
+            access_denied('legal_recycle_bin');
+        }
         if(!$id){
             set_alert('danger', _l('WrongEntry'));
             redirect(admin_url("Service/$ServID"));
@@ -234,7 +244,7 @@ class Cases_controller extends AdminController
 
         $data['selected_statuses'] = $selected_statuses;
 
-        if (has_permission('projects', '', 'view')) {
+        if (staff_can('view', 'projects')) {
             $selectedMember          = $appliedMember;
             $data['selectedMember']  = $selectedMember;
             $data['project_members'] = $this->case->get_distinct_projects_members();
@@ -250,7 +260,7 @@ class Cases_controller extends AdminController
 
     public function view($ServID,$id)
     {
-        if (has_permission('projects', '', 'view') || $this->case->is_member($id)) {
+        if (staff_can('view', 'projects') || $this->case->is_member($id)) {
             $slug = $this->legal->get_service_by_id($ServID)->row()->slug;
             $data['slug'] = $slug;
             close_setup_menu();
@@ -337,7 +347,7 @@ class Cases_controller extends AdminController
                 }
 
                 $__total_where_tasks = 'rel_type = "'.$slug.'" AND rel_id=' . $this->db->escape_str($id);
-                if (!has_permission('tasks', '', 'view')) {
+                if (!staff_can('view', 'tasks')) {
                     $__total_where_tasks .= ' AND ' . db_prefix() . 'tasks.id IN (SELECT taskid FROM ' . db_prefix() . 'task_assigned WHERE staffid = ' . get_staff_user_id() . ')';
 
                     if (get_option('show_all_tasks_for_project_member') == 1) {
@@ -481,7 +491,7 @@ class Cases_controller extends AdminController
 
             $other_projects_where .= ')';
 
-            if (!has_permission('projects', '', 'view')) {
+            if (!staff_can('view', 'projects')) {
                 $other_projects_where .= ' AND ' . db_prefix() . 'projects.id IN (SELECT project_id FROM ' . db_prefix() . 'my_members_cases WHERE staff_id=' . get_staff_user_id() . ')';
             }
 
@@ -506,7 +516,7 @@ class Cases_controller extends AdminController
         $success = false;
         $message = '';
         if ($this->input->is_ajax_request()) {
-            if (has_permission('projects', '', 'create') || has_permission('projects', '', 'edit')) {
+            if (staff_can('create', 'projects') || staff_can('edit', 'projects')) {
                 $status = get_case_status_by_id($this->input->post('status_id'));
 
                 $message = _l('project_marked_as_failed', $status['name']);
@@ -558,7 +568,7 @@ class Cases_controller extends AdminController
 
     public function download_all_files($ServID = '', $id)
     {
-        if ($this->case->is_member($id) || has_permission('projects', '', 'view')) {
+        if ($this->case->is_member($id) || staff_can('view', 'projects')) {
             $files = $this->case->get_files($id);
             if (count($files) == 0) {
                 set_alert('warning', _l('no_files_found'));
@@ -576,7 +586,7 @@ class Cases_controller extends AdminController
 
     public function export_project_data($ServID, $id)
     {
-        if (has_permission('projects', '', 'create')) {
+        if (staff_can('create', 'projects')) {
             app_pdf('case-data', LIBSPATH . 'pdf/Case_data_pdf',$ServID, $id);
         }
     }
@@ -603,7 +613,7 @@ class Cases_controller extends AdminController
 
     public function add_edit_members($ServID = '',$project_id)
     {
-        if (has_permission('projects', '', 'edit') || has_permission('projects', '', 'create')) {
+        if (staff_can('edit', 'projects')) {
             $this->case->add_edit_members($this->input->post(), $ServID, $project_id);
             redirect($_SERVER['HTTP_REFERER']);
         }
@@ -677,7 +687,7 @@ class Cases_controller extends AdminController
     public function delete_discussion($id)
     {
         $success = false;
-        if (has_permission('projects', '', 'delete')) {
+        if (staff_can('delete', 'projects')) {
             $success = $this->case->delete_discussion($id);
         }
         $alert_type = 'warning';
@@ -713,7 +723,7 @@ class Cases_controller extends AdminController
 
     public function change_activity_visibility($id, $visible)
     {
-        if (has_permission('projects', '', 'create')) {
+        if (staff_can('create', 'projects')) {
             if ($this->input->is_ajax_request()) {
                 $this->case->change_activity_visibility($id, $visible);
             }
@@ -768,7 +778,7 @@ class Cases_controller extends AdminController
 
     public function milestones($project_id, $ServID, $slug)
     {
-        if ($this->case->is_member($project_id) || has_permission('projects', '', 'view')) {
+        if ($this->case->is_member($project_id) || staff_can('view', 'projects')) {
             if ($this->input->is_ajax_request()) {
                 $this->app->get_table_data('milestones_case', [
                     'project_id' => $project_id,
@@ -785,11 +795,17 @@ class Cases_controller extends AdminController
             $message = '';
             $success = false;
             if (!$this->input->post('id')) {
+                if (!staff_can('create_milestones', 'projects')) {
+                    access_denied();
+                }
                 $id = $this->case->add_milestone($ServID, $this->input->post());
                 if ($id) {
                     set_alert('success', _l('added_successfully', _l('project_milestone')));
                 }
             } else {
+                if (!staff_can('edit_milestones', 'projects')) {
+                    access_denied();
+                }
                 $data = $this->input->post();
                 $id   = $data['id'];
                 unset($data['id']);
@@ -805,7 +821,7 @@ class Cases_controller extends AdminController
 
     public function delete_milestone($ServID='',$project_id, $id)
     {
-        if (has_permission('projects', '', 'delete')) {
+        if (staff_can('delete_milestones', 'projects')) {
             if ($this->case->delete_milestone($id)) {
                 set_alert('deleted', 'project_milestone');
             }
@@ -817,7 +833,7 @@ class Cases_controller extends AdminController
     {
         hooks()->do_action('before_do_bulk_action_for_project_files');
         $total_deleted       = 0;
-        $hasPermissionDelete = has_permission('projects', '', 'delete');
+        $hasPermissionDelete = staff_can('delete', 'projects');
         // bulk action for projects currently only have delete button
         if ($this->input->post()) {
             $fVisibility = $this->input->post('visible_to_customer') == 'true' ? 1 : 0;
@@ -839,7 +855,7 @@ class Cases_controller extends AdminController
 
     public function timesheets($project_id, $slug)
     {
-        if ($this->case->is_member($project_id) || has_permission('projects', '', 'view')) {
+        if ($this->case->is_member($project_id) || staff_can('view', 'projects')) {
             if ($this->input->is_ajax_request()) {
                 $this->app->get_table_data('timesheets_case', [
                     'project_id' => $project_id,
@@ -874,8 +890,8 @@ class Cases_controller extends AdminController
     {
         $assignees             = $this->tasks_model->get_task_assignees($task_id);
         $data                  = '';
-        $has_permission_edit   = has_permission('projects', '', 'edit');
-        $has_permission_create = has_permission('projects', '', 'edit');
+        $has_permission_edit   = staff_can('edit', 'projects');
+        $has_permission_create = staff_can('edit', 'projects');
         // The second condition if staff member edit their own timesheet
         if ($staff_id == 'undefined' || $staff_id != 'undefined' && (!$has_permission_edit || !$has_permission_create)) {
             $staff_id     = get_staff_user_id();
@@ -899,7 +915,7 @@ class Cases_controller extends AdminController
 
     public function remove_team_member($ServID,$project_id, $staff_id)
     {
-        if (has_permission('projects', '', 'edit') || has_permission('projects', '', 'create')) {
+        if (staff_can('edit', 'projects')) {
             if ($this->case->remove_team_member($ServID,$project_id, $staff_id)) {
                 set_alert('success', _l('project_member_removed'));
             }
@@ -920,7 +936,7 @@ class Cases_controller extends AdminController
 
     public function copy($ServID,$project_id)
     {
-        if (has_permission('projects', '', 'create')) {
+        if (staff_can('create', 'projects')) {
             $id = $this->case->copy($ServID,$project_id, $this->input->post());
             if ($id) {
                 set_alert('success', _l('project_copied_successfully'));
@@ -952,7 +968,7 @@ class Cases_controller extends AdminController
 
     public function mass_stop_timers($project_id, $billable = 'false')
     {
-        if (has_permission('invoices', '', 'create')) {
+        if (staff_can('create', 'invoices')) {
             $where = [
                 'billed'       => 0,
                 'startdate <=' => date('Y-m-d'),
@@ -984,7 +1000,7 @@ class Cases_controller extends AdminController
 
     public function get_pre_invoice_project_info($ServID, $project_id)
     {
-        if (has_permission('invoices', '', 'create')) {
+        if (staff_can('create', 'invoices')) {
             $slug = $this->legal->get_service_by_id($ServID)->row()->slug;
             $data['billable_tasks'] = $this->case->get_tasks($project_id, [
                 'billable'     => 1,
@@ -1016,7 +1032,7 @@ class Cases_controller extends AdminController
 
     public function get_invoice_project_data($ServID)
     {
-        if (has_permission('invoices', '', 'create')) {
+        if (staff_can('create', 'invoices')) {
             $type       = $this->input->post('type');
             $project_id = $this->input->post('project_id');
             // Check for all cases
@@ -1223,7 +1239,7 @@ class Cases_controller extends AdminController
 
     public function invoice_project($ServID, $project_id)
     {
-        if (has_permission('invoices', '', 'create')) {
+        if (staff_can('create', 'invoices')) {
             $slug = $this->legal->get_service_by_id($ServID)->row()->slug;
             $this->load->model('invoices_model');
             $data               = $this->input->post();
